Move autocast from core.transforms to transforms
