+++ thunder/common.py
@@ -40,7 +40,7 @@ import thunder.distributed as dist
 import thunder.torch as ltorch  # we need to import thunder.torch before the below for import ordering...
 from thunder.extend import Executor, get_default_executors, get_always_executors, OperatorExecutor, add_executor_lists
 import thunder.executors as executors
-from thunder.core.transforms import autocast
+from thunder.transforms import autocast
 from thunder.core.dtypes import to_dtype
 
 import torch as torch
