+++ thunder/tests/test_core.py
@@ -2172,7 +2172,7 @@ def test_inplace(executor, device, _):
 
 @instantiate(dtypes=NOTHING)
 def test_thunder_autocast_transform(executor, device, _):
-    from thunder.core.transforms import autocast
+    from thunder.transforms import autocast
 
     def f(a, b, c):
         return a @ (b + c)
